backend/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
bcrypt==4.0.1
beautifulsoup4==4.12.2
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.1.8
cryptography==46.0.3
dnspython==2.7.0
ecdsa==0.19.1
email-validator==2.3.0
exceptiongroup==1.3.1
fastapi==0.109.0
greenlet==3.3.0
h11==0.16.0
httptools==0.7.1
idna==3.11
lxml>=5.0.0
passlib==1.7.4
playwright
psycopg2-binary==2.9.11
pyasn1==0.6.1
pycparser==2.23
pydantic==2.12.5

pyee==11.0.1
pyOpenSSL==25.3.0
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0.3
requests==2.31.0
rsa==4.9.1
signxml
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy>=2.0.30
starlette==0.35.1
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.27.0

watchfiles==1.1.1
websockets==15.0.1

# Testing dependencies
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.23.2
httpx==0.25.2

